Implemented portion of removing LaneSet direct replacements
Currently, there are a few helper functions to help with this transition
I want all protected tick logic (tick0Immune) to be WITHIN LaneSet<T>, and I need to remove all reassignments to do so. Saving here temporarily because I now have to edit moving logic and moving logic is a pain.